Heat/Radiation v2 part 1 - REACTORS BROKEN
Switching to a new system that generates neutrons based on fuel, splits
neutrons into fast/slow based on heat, then generates heat/power based
on neutron propagation.
Adding "incident radiation" so we get network effects from fuel rods
placed near each other.
Rebalancing lots of numbers.
Changing TileEntityFuelRod to be a "proxy" object for control rods
Commenting out some code in the reactor temporarily. REACTORS ARE
TEMPORARILY BROKEN.
Switching radiation to continuous system instead of discrete (doubles
instead of ints)
Removing unused Fuel Rod renderer now that the sweet new Control Rod
renderer has supplanted it

+++ erogenousbeef/bigreactors/api/IRadiationPulse.java
@@ -1,15 +1,15 @@
 package erogenousbeef.bigreactors.api;
 
 public interface IRadiationPulse {
-	public int getFastRadiation();
-	public int getSlowRadiation();
-	public int getPowerProduced();
+	public double getFastRadiation();
+	public double getSlowRadiation();
+	public double getPowerProduced();
 	public double getHeatProduced();
 	
 	public int getTimeToLive();
 	
-	public void setFastRadiation(int newValue);
-	public void setSlowRadiation(int newValue);
+	public void setFastRadiation(double newValue);
+	public void setSlowRadiation(double newValue);
 	public void addPower(double d);
 	public void changeHeat(double difference);
